refactor: extract UsersSection component
